requirements.txt
numpy
faiss-cpu
openai